Machine Learning with MLflow/requirements.txt
ensure
fastapi
joblib
mlflow
pandas
python-box
python-dotenv
PyYAML
scikit-learn
uvicorn